implementing ocr

+++ archive/models.py
@@ -1,44 +1,62 @@
 import os
+import re
 import boto3
 import pytesseract
+from subprocess import call
 from PIL import Image
+from lxml import html
+from django.db import models
+from datetime import datetime
 from PyPDF2 import PdfFileMerger
 from django.conf import settings
 from django.urls import reverse
-from django.db import models
 from django.core.files import File
 from tempfile import TemporaryFile
 
+
 class ArchivedFileModel(models.Model):
+    """
+    Base model for pages and issues.
+    """
 
-    def local_path(self, filename):
+    def local_path(self, path):
         return os.path.join(
-            settings.PROCESSED_DIR, filename)
-
-    def upload(self, path):
-        s3 = boto3.resource('s3')
-        bucket = s3.Bucket(settings.ARCHIVE_BUCKET_NAME)
-        local_path = self.local_path(path)
-        bucket.upload_file(
-            local_path,
-            path,
-            ExtraArgs={"ACL": "public-read"}
+            settings.PROCESSED_DIR,
+            path
         )
+
+    def upload_file(self, path):
+        """
+        Upload to the archival bucket.
+        """
+        local_path = self.local_path(path)
+        if os.path.exists(self.local_path):
+            s3 = boto3.resource('s3')
+            bucket = s3.Bucket(settings.ARCHIVE_BUCKET_NAME)
+            bucket.upload_file(
+                local_path,
+                path,
+                ExtraArgs={'ACL': 'public-read'}
+            )
+        else:
+            raise Exception('Local file not found.')
 
     class Meta:
         abstract = True
 
 
 class Page(ArchivedFileModel):
-    """A single page from an issue."""
-
+    """
+    A single page from an issue.
+    """
     date = models.DateField()
     page_number = models.IntegerField()
     text = models.TextField()
     issue = models.ForeignKey('Issue', null=True, related_name='pages')
+    processed_datetime = models.DateTimeField(null=True)
 
     @property
-    def path(self):
+    def filename(self):
         return '{0}/{1}/{2}/dailycal_{0}{1}{2}_{3}'.format(
             self.date.year,
             self.date.strftime('%m'),
@@ -46,81 +64,123 @@ class Page(ArchivedFileModel):
             format(self.page_number,'02'))
 
     @property
-    def tif(self):
-        return self.path + '.tif'
+    def tif_path(self):
+        return self.filename + '.tif'
 
     @property
-    def local_tif_path(self):
-        return self.local_path(self.tif)
-
-    def has_tif(self):
-        return os.path.exists(self.local_tif_path)
-
-    def upload_tif(self):
-        if self.has_tif:
-            self.upload(self.tif)
-        else:
-            raise Exception('No local TIF file!')
-
-    @property
-    def jpg(self):
-        return self.path + '.jpg'
-
-    @property
-    def local_jpg_path(self):
-        return self.local_path(self.jpg)
-
-    def has_jpg(self):
-        return os.path.exists(self.local_jpg_path)
+    def jpg_path(self):
+        return self.filename + '.jpg'
 
     def save_jpg(self):
-        if self.has_tif:
-            tif_file = Image.open(self.local_tif_path)
-            with open(self.local_jpg_path, 'wb') as f:
+        """
+        Save a JPG version of this page.
+        """
+        # Check that the tif exists locally
+        local_tif_path = self.local_path(self.tif_path)
+        if os.path.exists(local_tif_path):
+            tif_file = Image.open(local_tif_path)
+            # Save a JPG version
+            local_jpg_path = self.local_path(self.jpg_path)
+            with open(local_jpg_path, 'wb') as f:
                 tif_file.save(f, 'JPEG')
         else:
-            raise Exception('No tif file to generate JPG from!')
-
-    def upload_jpg(self):
-        if self.has_jpg:
-            self.upload(self.jpg)
-        else:
-            raise Exception('No local JPG file!')
+            raise Exception('No TIF to generate JPG from.')
     
     @property
-    def pdf(self):
-        return self.path + '.pdf'
-
-    @property
-    def local_pdf_path(self):
-        return self.local_path(self.pdf)
-
-    def has_pdf(self):
-        return os.path.exists(self.local_pdf_path)
+    def pdf_path(self):
+        return self.filename + '.pdf'
 
     def save_pdf(self):
-        if self.has_tif:
-            tif_file = Image.open(self.local_tif_path)
-            with open(self.local_pdf_path, 'wb') as f:
+        """
+        Save a PDF version of this page.
+        """
+        # Check that the tif exists locally
+        local_tif_path = self.local_path(self.tif_path)
+        if os.path.exists(local_tif_path):
+            tif_file = Image.open(local_tif_path)
+            # Save a PDF version
+            local_pdf_path = self.local_path(self.pdf_path)
+            with open(local_pdf_path, 'wb') as f:
                 tif_file.save(f, 'PDF')
         else:
-            raise Exception('No tif file to generate PDF from!')
+            raise Exception('No TIF file to generate PDF from.')
 
-    def upload_pdf(self):
-        if self.has_pdf:
-            self.upload(self.pdf)
-        else:
-            raise Exception('No local PDF file!')
+    def process(self):
+        """
+        Generate JPG and PDF versions from the raw image. Run the raw image
+        through tesseract to generate an hOCR file, and import the text
+        from that file into the database.
+        """
+        self.save_jpg()
+        self.save_pdf()
+        self.save_hocr_file()
+        self.save_ocr_text()
+        self.processed_datetime = datetime.now()
+        self.save()
+
+    def upload(self):
+        """
+        Upload TIF, JPG, hOCR and PDF.
+        """
+        self.upload_file(self.local_path(self.tif_path))
+        self.upload_file(self.local_path(self.jpg_path))
+        self.upload_file(self.local_path(self.hocr_path))
+        self.upload_file(self.local_path(self.pdf_path))
+
+    @property
+    def hocr_path(self):
+        return self.filename + '.hocr'
+
+    def save_hocr_file(self):
+        """
+        Generate hOCR file using tesseract.
+        """
+        tif_path = self.local_path(self.tif_path)
+        call([
+            'tesseract',
+            tif_path,
+            self.local_path(self.filename),
+            '-c language_model_penalty_non_freq_dict_word=4',
+            '-c language_model_penalty_non_dict_word=2',
+            'hocr',
+        ])
+
+    def save_ocr_text(self):
+        """
+        Save raw hOCR text to database. Adapted from
+        https://github.com/tmbdev/hocr-tools/blob/master/hocr-lines
+        """
+        hocr_path = self.local_path(self.filename + '.hocr')
+        
+        if not os.path.exists(hocr_path):
+            raise Exception('This page does not have an hOCR file.')
+
+        text = ''
+
+        doc = html.parse(hocr_path)
+        lines = doc.xpath("//*[@class='ocr_line']")
+
+        for line in lines:
+            textnodes = line.xpath(".//text()")
+            s = ''.join([text for text in textnodes])
+            text += re.sub(r'\s+',' ',s)
+            text += ' '
+
+        self.text = text
+        self.save()
 
     def __str__(self):
         return '{}: p. {}'.format(self.date, self.page_number)
 
     class Meta:
-        ordering = ['-date','page_number']
+        ordering = ['date','page_number']
         unique_together = (('date', 'page_number'),)
 
 
 class Issue(ArchivedFileModel):
+    """
+    A single issue published on a particular day.
+    """
     date = models.DateField(unique=True)
     sponsor = models.CharField(
         max_length=200,
@@ -131,11 +191,20 @@ class Issue(ArchivedFileModel):
 
     @property
     def date_parts_list(self):
-        """List of year, month, day."""
-        return [self.date.strftime('%Y'), self.date.strftime('%m'), self.date.strftime('%d')]
+        """
+        List of [year, month, day].
+        """
+        return [
+            self.date.strftime('%Y'),
+            self.date.strftime('%m'),
+            self.date.strftime('%d')
+        ]
 
     @property
     def date_parts_dict(self):
+        """
+        Dict of {"year": year, "month": month, "day": day}.
+        """
         return {
             'year': self.date.strftime('%Y'),
             'month': self.date.strftime('%m'),
@@ -147,57 +216,52 @@ class Issue(ArchivedFileModel):
         return '{0}/{1}/{2}'.format(*self.date_parts_list)
 
     @property
-    def path(self):
-        return '{0}/{1}/{2}/dailycal_{0}{1}{2}_issue'.format(*self.date_parts_list)
+    def filename(self):
+        return '{0}/{1}/{2}/dailycal_{0}{1}{2}_issue'.format(
+            *self.date_parts_list
+        )
 
     @property
-    def pdf(self):
-        return self.path + '.pdf'
-
-    @property
-    def local_pdf_path(self):
-        return self.local_path(self.pdf)
-
-    def has_pdf(self):
-        return os.path.exists(self.local_pdf_path)
+    def pdf_path(self):
+        return self.filename + '.pdf'
 
     def save_pdf(self):
+        """
+        Generate the issue PDF. Assumes all page PDFs exist locally.
+        """
+
         outfile = PdfFileMerger()
+
+        # Add PDF for each page in the issue
         for page in self.pages.all():
-            # If we don't already have the PDF
-            if not page.has_pdf:
-                # Try to create it
-                if page.has_tif:
-                    page.save_pdf()
-                # Or download it
-                else:
-                    bucket.download_file(page.pdf, page.local_pdf_path)
-            outfile.append(page.local_pdf_path)
+            local_pdf_path = page.local_path(page.pdf_path)
+            outfile.append(local_pdf_path)
 
         # Write the issue PDF
-        issue_path = os.path.join(settings.PROCESSED_DIR, self.pdf)
-        issue_pdf = open(issue_path, 'w+')
-        outfile.write(issue_path)
-
-    def upload_pdf(self):
-        if self.has_pdf:
-            self.upload(self.pdf)
-        else:
-            raise Exception('No local PDF file!')
+        local_pdf_path = self.local_path(self.pdf_path)
+        issue_pdf = open(local_pdf_path, 'w+')
+        outfile.write(local_pdf_path)
 
     def process(self):
+        """
+        Process pages and generate PDF.
+        """
         for page in self.pages.all():
-            if page.has_tif:
-                page.upload_tif()
-                page.save_jpg()
-                page.upload_jpg()
-                page.save_pdf()
-                page.upload_pdf()
+            page.process()
+
         self.save_pdf()
-        self.upload_pdf()
+
+    def upload(self):
+        """
+        Upload pages and issue PDF.
+        """
+        for page in self.pages.all():
+            page.upload()
+
+        self.upload_file(self.local_path(self.pdf_path))
 
     def __str__(self):
         return '{}'.format(self.date)
 
     class Meta:
-        ordering = ['-date']
+        ordering = ['date']
